Add libraries support for python 3
Introduces a simple mechanism to set cflags/ldflags for some optional libs that can be linked with python. This mechanism is created in order to keep the code clean and readable, and should make easier to add other optional libs. The supported libraries are:

  - sqlite3
  - openssl
  - libexpat
  - libffi

The libraries should be tested with a real app and further changes may be needed. Also add termios patch in order to avoid runtime error when running the app.