Fix applied patch from tito's
The previously applied patch 2d54514 leads into merge conflicts. I don't resolve well the patch because I keep the old changes coming from be278a3 (which leads into error, and the new changes, which wasn't applied, seems to not fail, but maybe we should also add the hostpython3cristax recipe...so...added in this commit).

Note: The problematic code has been comented and a todo message added to review that.