Fix cflags/ldflags for mk files
Since the new python recipes build system were introduced some of the bootstraps needs to be updated because the cflags/ldflags still points to the old python location. This will make easier to add compatibility to the new python's build system for the non sdl2 bootstraps, but be aware that some more jobs has to be done in order to make it work those bootstraps...so...this prepares the terrain to fix it.