Fix bootstraps and add python3's compatibility
This commit should fix the bootstraps (service_only and webview). The pygame bootstrap is not covered here because it depends of some dependencies that makes it incompatible with the new python build system...so it will be treated later as a special case. Those changes are almost the same than the ones applied by @inclement into the sdl2 bootstrap with a few additions of my own.

Note: This changes were not tested in an app...so...we should make sure that works as expected