Reintroduces the original python2's recipe by @tito as python2legacy
(This branch and commit  has been created as a sample to help decide if it is worth it to maintain the old python2 recipe as a new recipe named python2legacy. here we will see the minimal changes that we should make to make it work.)

In order to maintain compatibility to pygame bootstrap and old sdl libraries, we perform several operations at once, this way can be reverted easily:
  - add python2legacy compatibility to basic recipes needed for pygame and bootstraps
  - solve the conflicts between python recipes
  - cause the pygame bootstrap has troubles to successfully build with the new python2's build, we enforce to build with python2legacy
  - update the core files to the new python2legacy
  - add compatibility python2legacy to all bootstraps
  - add some python2legacy tests

Note: this new addition should be properly documented and add more recipes compatibility for python2legacy.

References: Pull Request #1460