Refactor hostpython3's recipe into a new base class HostPythonRecipe ... ... ... Add ability to hide some log messages and hide the unpack logs
This should allow us to reduce console output when decompressing files (most of the times we don't need that list) and also will reduce the probabilities that travis fails due to the 4mb logs limit. This will fix the travis error when uncompress "the boost recipe" files, which has a lot of files and overpass the travis log limit.